feat(junit5): adjust junit5-extension
Related to CAM-11955
