+++ pom.xml
@@ -80,6 +80,7 @@
         <module>typed-values</module>
         <module>engine-dmn</module>
         <module>test-utils/testcontainers</module>
+        <module>test-utils/junit5-extension</module>
 
         <module>engine</module>
         <module>engine-cdi</module>
@@ -275,6 +276,8 @@
         <module>distro/webjar</module>
 
         <module>clients/java</module>
+
+        <module>test-utils/junit5-extension</module>
 
         <module>internal-dependencies</module>
         <module>database</module>
