+++ test-utils/junit5-extension/pom.xml
@@ -3,20 +3,22 @@
   <artifactId>camunda-bpm-junit5</artifactId>
   <packaging>jar</packaging>
   
-  <name>Camunda BPM JUnit 5 Extension</name>
+  <name>Camunda Platform - JUnit 5 Extension</name>
   
   <parent>
-    <groupId>org.camunda.bpm.extension</groupId>
-    <artifactId>camunda-bpm-junit5-parent</artifactId>
-    <version>1.1.1-SNAPSHOT</version>
+    <groupId>org.camunda.bpm</groupId>
+    <artifactId>camunda-database-settings</artifactId>
+    <relativePath>../../database</relativePath>
+    <version>7.17.0-SNAPSHOT</version>
   </parent>
   
   <dependencyManagement>
     <dependencies>
       <dependency>
         <groupId>org.camunda.bpm</groupId>
-        <artifactId>camunda-bom</artifactId>
-        <version>7.16.0</version>
+        <artifactId>camunda-core-internal-dependencies</artifactId>
+        <version>${project.version}</version>
+        <scope>import</scope>
         <type>pom</type>
       </dependency>
     </dependencies>
@@ -26,7 +28,6 @@
     <dependency>
       <groupId>org.assertj</groupId>
       <artifactId>assertj-core</artifactId>
-      <version>3.22.0</version>
       <scope>test</scope>
     </dependency>
     <dependency>
@@ -38,38 +39,17 @@
     <dependency>
       <groupId>org.camunda.bpm</groupId>
       <artifactId>camunda-engine</artifactId>
-      <version>7.16.0</version>
       <scope>provided</scope>
     </dependency>
     <dependency>
       <groupId>com.h2database</groupId>
       <artifactId>h2</artifactId>
-      <version>1.4.200</version>
       <scope>test</scope>
     </dependency>
     <dependency>
       <groupId>ch.qos.logback</groupId>
       <artifactId>logback-classic</artifactId>
-      <version>1.2.10</version>
       <scope>test</scope>
     </dependency>
   </dependencies>
-
-  <build>
-    <plugins>
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-surefire-plugin</artifactId>
-        <version>3.0.0-M5</version>
-      </plugin>
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-javadoc-plugin</artifactId>
-        <configuration>
-          <source>8</source>
-        </configuration>
-      </plugin>
-      
-    </plugins>
-  </build>
 </project>
