+++ test-utils/junit5-extension/src/main/java/org/camunda/bpm/engine/test/junit5/ProcessEngineExtension.java
@@ -1,0 +1,209 @@
+/*
+ * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH
+ * under one or more contributor license agreements. See the NOTICE file
+ * distributed with this work for additional information regarding copyright
+ * ownership. Camunda licenses this file to you under the Apache License,
+ * Version 2.0; you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.camunda.bpm.engine.test.junit5;
+
+import java.lang.reflect.Field;
+import java.lang.reflect.Method;
+import java.util.Arrays;
+import java.util.function.Supplier;
+
+import org.camunda.bpm.engine.ProcessEngine;
+import org.camunda.bpm.engine.RepositoryService;
+import org.camunda.bpm.engine.impl.ProcessEngineLogger;
+import org.camunda.bpm.engine.impl.cfg.ProcessEngineConfigurationImpl;
+import org.camunda.bpm.engine.impl.test.TestHelper;
+import org.camunda.bpm.engine.test.Deployment;
+import org.camunda.bpm.engine.test.RequiredHistoryLevel;
+import org.junit.jupiter.api.Assumptions;
+import org.junit.jupiter.api.extension.AfterTestExecutionCallback;
+import org.junit.jupiter.api.extension.BeforeTestExecutionCallback;
+import org.junit.jupiter.api.extension.ExtensionContext;
+import org.junit.jupiter.api.extension.ParameterContext;
+import org.junit.jupiter.api.extension.ParameterResolutionException;
+import org.junit.jupiter.api.extension.ParameterResolver;
+import org.junit.jupiter.api.extension.TestInstancePostProcessor;
+import org.junit.jupiter.api.extension.TestWatcher;
+import org.slf4j.Logger;
+
+/**
+ * Junit 5 Extension to create and inject a {@link ProcessEngine} into the test class.
+ * <p>
+ * If you provide a {@code camunda.cfg.xml} file on the classpath. This file is used to configure the process engine.
+ * <br>
+ * Usage:
+ * </p>
+ * <pre>
+ * <code>@ExtendWith(ProcessEngineExtension.class)</code>
+ * public class YourTest {
+ *
+ *   // provide a property where the extension can inject the process engine
+ *   public ProcessEngine processEngine;
+ *
+ *   ...
+ * }
+ * </pre>
+ *
+ * <p>
+ * If you want to choose the {@code camunda.cfg.xml} file that is used in the test programmatically,
+ * you can register the extension directly and use the builder pattern to configure it.
+ * <br>
+ * Usage with configuration:
+ *
+ * </p>
+ * Usage:
+ * <p>
+ * <pre>
+ * <code>@RegisterExtension</code>
+ * ProcessEngineExtension extension = ProcessEngineExtension.builder()
+ *    .configurationResource("myConfigurationFile.xml")
+ *    .build();}
+ * </pre>
+ * </p>
+ * <p>
+ * You can declare a deployment with the {@link Deployment} annotation. This
+ * base class will make sure that this deployment gets deployed before the setUp
+ * and {@link RepositoryService#deleteDeployment(String, boolean) cascade
+ * deleted} after the tearDown.
+ * </p>
+ * <p>
+ * If you need the history service for your tests then you can specify the
+ * required history level of the test method or class, using the
+ * {@link RequiredHistoryLevel} annotation. If the current history level of the
+ * process engine is lower than the specified one then the test is skipped.
+ * </p>
+ */
+public class ProcessEngineExtension implements TestWatcher,
+    TestInstancePostProcessor, BeforeTestExecutionCallback, AfterTestExecutionCallback,
+    ParameterResolver {
+
+  protected static Supplier<IllegalStateException> illegalStateException(String msg) {
+    return () -> new IllegalStateException(msg);
+  }
+
+  protected static final Logger LOG = ProcessEngineLogger.TEST_LOGGER.getLogger();
+  protected ProcessEngine processEngine;
+  protected ProcessEngineConfigurationImpl processEngineConfiguration;
+
+  protected String configurationResource = "camunda.cfg.xml";
+
+  protected String deploymentId;
+
+  public static ProcessEngineExtension builder() {
+    return new ProcessEngineExtension();
+  }
+
+  public ProcessEngineExtension configurationResource(String configurationResource) {
+    this.configurationResource = configurationResource;
+    return this;
+  }
+
+  public ProcessEngineExtension useProcessEngine(ProcessEngine engine) {
+    this.setProcessEngine(engine);
+    return this;
+  }
+
+  public ProcessEngine getProcessEngine() {
+    return processEngine;
+  }
+
+  public void setProcessEngine(ProcessEngine processEngine) {
+    this.processEngine = processEngine;
+  }
+
+  public String getConfigurationResource() {
+    return configurationResource;
+  }
+
+  public ProcessEngineConfigurationImpl getProcessEngineConfiguration() {
+    return processEngineConfiguration;
+  }
+
+  public String getDeploymentId() {
+    return deploymentId;
+  }
+
+  @Override
+  public void postProcessTestInstance(Object testInstance, ExtensionContext context) {
+    if (processEngine == null) {
+      initializeProcessEngine();
+    }
+    Arrays.stream(testInstance.getClass().getDeclaredFields())
+      .filter(field -> field.getType() == ProcessEngine.class)
+      .forEach(field -> inject(testInstance, field));
+  }
+
+  protected void initializeProcessEngine() {
+    processEngine = TestHelper.getProcessEngine(configurationResource);
+    processEngineConfiguration = (ProcessEngineConfigurationImpl) processEngine.getProcessEngineConfiguration();
+  }
+
+  protected void inject(Object instance, Field field) {
+    field.setAccessible(true);
+    try {
+      field.set(instance, processEngine);
+    } catch (IllegalAccessException iae) {
+      throw new RuntimeException(iae);
+    }
+  }
+
+  @Override
+  public void beforeTestExecution(ExtensionContext context) {
+    LOG.debug("beforeTestExecution: {}", context.getDisplayName());
+
+    final Method testMethod = context.getTestMethod().orElseThrow(illegalStateException("testMethod not set"));
+    final Class<?> testClass = context.getTestClass().orElseThrow(illegalStateException("testClass not set"));
+
+    deploymentId = TestHelper.annotationDeploymentSetUp(processEngine, testClass, testMethod.getName(), null, testMethod.getParameterTypes());
+    boolean hasRequiredHistoryLevel = TestHelper.annotationRequiredHistoryLevelCheck(processEngine, testClass, testMethod.getName(), testMethod.getParameterTypes());
+
+    Assumptions.assumeTrue(hasRequiredHistoryLevel, "ignored because the current history level is too low");
+  }
+
+  @Override
+  public void afterTestExecution(ExtensionContext context) {
+    final String testMethod = context.getTestMethod().orElseThrow(illegalStateException("testMethod not set")).getName();
+    final Class<?> testClass = context.getTestClass().orElseThrow(illegalStateException("testClass not set"));
+
+   TestHelper.annotationDeploymentTearDown(processEngine, deploymentId, testClass, testMethod);
+   deploymentId = null;
+  }
+
+  @Override
+  public boolean supportsParameter(ParameterContext parameterContext, ExtensionContext extensionContext)
+      throws ParameterResolutionException {
+    return parameterContext.getParameter().getType().equals(ProcessEngine.class);
+  }
+
+  @Override
+  public Object resolveParameter(ParameterContext parameterContext, ExtensionContext extensionContext)
+      throws ParameterResolutionException {
+    if (ProcessEngine.class.equals(parameterContext.getParameter().getType())) {
+      LOG.debug("resolve the processEngine as parameter");
+      return getProcessEngine();
+    } else {
+      return null;
+    }
+  }
+
+  public ProcessEngineExtension build() {
+    if (processEngine == null) {
+      initializeProcessEngine();
+    }
+    return this;
+  }
+
+}
