+++ test-utils/junit5-extension/src/test/java/org/camunda/bpm/engine/test/junit5/ProcessEngineExtensionJunit5Test.java
@@ -1,4 +1,20 @@
-package org.camunda.bpm.extension.junit5;
+/*
+ * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH
+ * under one or more contributor license agreements. See the NOTICE file
+ * distributed with this work for additional information regarding copyright
+ * ownership. Camunda licenses this file to you under the Apache License,
+ * Version 2.0; you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.camunda.bpm.engine.test.junit5;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -10,15 +26,14 @@ import org.camunda.bpm.engine.TaskService;
 import org.camunda.bpm.engine.task.Task;
 import org.camunda.bpm.engine.test.Deployment;
 import org.camunda.bpm.engine.test.RequiredHistoryLevel;
-import org.camunda.bpm.extension.junit5.test.ProcessEngineExtension;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 
 @ExtendWith(ProcessEngineExtension.class)
 public class ProcessEngineExtensionJunit5Test {
-  
+
   ProcessEngine engine;
-  
+
   @Test
   @Deployment
   public void extensionUsageExample() {
@@ -32,7 +47,7 @@ public class ProcessEngineExtensionJunit5Test {
     taskService.complete(task.getId());
     assertEquals(0, runtimeService.createProcessInstanceQuery().count());
   }
-  
+
   /**
    * The extension should work with tests that have no deployment annotation
    */
@@ -44,7 +59,7 @@ public class ProcessEngineExtensionJunit5Test {
   @Test
   @RequiredHistoryLevel(ProcessEngineConfiguration.HISTORY_AUDIT)
   public void requiredHistoryLevelAudit() {
-    
+
     assertThat(currentHistoryLevel()).isIn(
         ProcessEngineConfiguration.HISTORY_AUDIT, ProcessEngineConfiguration.HISTORY_FULL);
   }
@@ -54,7 +69,7 @@ public class ProcessEngineExtensionJunit5Test {
   public void requiredHistoryLevelActivity() {
 
     assertThat(currentHistoryLevel()).isIn(
-        ProcessEngineConfiguration.HISTORY_ACTIVITY, 
+        ProcessEngineConfiguration.HISTORY_ACTIVITY,
         ProcessEngineConfiguration.HISTORY_AUDIT,
         ProcessEngineConfiguration.HISTORY_FULL);
   }
