+++ test-utils/junit5-extension/src/test/java/org/camunda/bpm/engine/test/junit5/ProcessEngineExtensionParameterizedJunit5Test.java
@@ -1,4 +1,20 @@
-package org.camunda.bpm.extension.junit5;
+/*
+ * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH
+ * under one or more contributor license agreements. See the NOTICE file
+ * distributed with this work for additional information regarding copyright
+ * ownership. Camunda licenses this file to you under the Apache License,
+ * Version 2.0; you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.camunda.bpm.engine.test.junit5;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -13,7 +29,6 @@ import org.camunda.bpm.engine.TaskService;
 import org.camunda.bpm.engine.history.HistoricVariableInstance;
 import org.camunda.bpm.engine.task.Task;
 import org.camunda.bpm.engine.test.Deployment;
-import org.camunda.bpm.extension.junit5.test.ProcessEngineExtension;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EmptySource;
@@ -21,9 +36,9 @@ import org.junit.jupiter.params.provider.ValueSource;
 
 @ExtendWith(ProcessEngineExtension.class)
 public class ProcessEngineExtensionParameterizedJunit5Test {
-  
+
   ProcessEngine engine;
-  
+
   @ParameterizedTest
   @ValueSource(ints = {1, 2})
   @Deployment
@@ -41,7 +56,7 @@ public class ProcessEngineExtensionParameterizedJunit5Test {
 
   @ParameterizedTest
   @ValueSource(strings = {"A", "B"})
-  @Deployment(resources = "org/camunda/bpm/extension/junit5/ProcessEngineExtensionParameterizedJunit5Test.extensionUsageExample.bpmn20.xml")
+  @Deployment(resources = "org/camunda/bpm/engine/test/junit5/ProcessEngineExtensionParameterizedJunit5Test.extensionUsageExample.bpmn20.xml")
   public void extensionUsageExampleWithNamedAnnotation(String value) {
     Map<String,Object> variables = new HashMap<>();
     variables.put("key", value);
@@ -54,7 +69,7 @@ public class ProcessEngineExtensionParameterizedJunit5Test {
 
     taskService.complete(task.getId());
     assertEquals(0, runtimeService.createProcessInstanceQuery().count());
-    
+
     HistoryService historyService = engine.getHistoryService();
     HistoricVariableInstance variableInstance = historyService.createHistoricVariableInstanceQuery().singleResult();
     assertThat(variableInstance.getValue()).isEqualTo(value);
