+++ test-utils/junit5-extension/src/test/java/org/camunda/bpm/engine/test/junit5/deployment/ProcessEngineExtensionDeploymentIdAccess.java
@@ -1,18 +1,31 @@
-package org.camunda.bpm.extension.junit5.deployment;
+/*
+ * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH
+ * under one or more contributor license agreements. See the NOTICE file
+ * distributed with this work for additional information regarding copyright
+ * ownership. Camunda licenses this file to you under the Apache License,
+ * Version 2.0; you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.camunda.bpm.engine.test.junit5.deployment;
 
-import org.camunda.bpm.extension.junit5.test.ProcessEngineExtension;
+import org.camunda.bpm.engine.test.junit5.ProcessEngineExtension;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtensionContext;
 
-/**
- * @author Martin Schimak
- */
 public class ProcessEngineExtensionDeploymentIdAccess {
 
   @Test
   public void testDeploymentIdWriteableForExtensions() {
 
-    class ProcessEngineExtensionExtension extends org.camunda.bpm.extension.junit5.test.ProcessEngineExtension {
+    class ProcessEngineExtensionExtension extends ProcessEngineExtension {
 
       @Override
       public void beforeTestExecution(ExtensionContext context) {
@@ -30,7 +43,7 @@ public class ProcessEngineExtensionDeploymentIdAccess {
   @Test
   public void testDeploymentIdReadableForExtensionsAndWrappers() {
 
-    class ProcessEngineExtensionExtension extends org.camunda.bpm.extension.junit5.test.ProcessEngineExtension {
+    class ProcessEngineExtensionExtension extends ProcessEngineExtension {
 
       @Override
       public void afterTestExecution(ExtensionContext context) {
