+++ test-utils/junit5-extension/src/test/java/org/camunda/bpm/engine/test/junit5/registerExtension/RegisterNewProcessEngineTest.java
@@ -1,4 +1,20 @@
-package org.camunda.bpm.extension.junit5.registerExtension;
+/*
+ * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH
+ * under one or more contributor license agreements. See the NOTICE file
+ * distributed with this work for additional information regarding copyright
+ * ownership. Camunda licenses this file to you under the Apache License,
+ * Version 2.0; you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.camunda.bpm.engine.test.junit5.registerExtension;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
@@ -7,31 +23,31 @@ import org.camunda.bpm.engine.ProcessEngineConfiguration;
 import org.camunda.bpm.engine.RuntimeService;
 import org.camunda.bpm.engine.TaskService;
 import org.camunda.bpm.engine.test.Deployment;
-import org.camunda.bpm.extension.junit5.test.ProcessEngineExtension;
+import org.camunda.bpm.engine.test.junit5.ProcessEngineExtension;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
 
 public class RegisterNewProcessEngineTest {
-  
+
   private static ProcessEngine testEngine = ProcessEngineConfiguration
       .createStandaloneInMemProcessEngineConfiguration()
       .setProcessEngineName("testEngine")
-      // Use a new database to resolve the conflict with existing 
+      // Use a new database to resolve the conflict with existing
       // in-memory-database. The tables will be removed after the test.
       .setJdbcUrl("jdbc:h2:mem:camunda-test")
       .buildProcessEngine();
-  
+
   @RegisterExtension
   ProcessEngineExtension extension = ProcessEngineExtension.builder()
       .useProcessEngine(testEngine)
       .build();
-  
+
   @AfterAll
   public static void closeProcessEngine() {
     testEngine.close();
   }
-  
+
   @Test
   @Deployment(resources = "processes/subProcess.bpmn")
   public void testUseProcessEngine() {
