+++ test-utils/junit5-extension/src/test/java/org/camunda/bpm/engine/test/junit5/registerExtension/RegisterProcessEngineExtensionTest.java
@@ -1,4 +1,20 @@
-package org.camunda.bpm.extension.junit5.registerExtension;
+/*
+ * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH
+ * under one or more contributor license agreements. See the NOTICE file
+ * distributed with this work for additional information regarding copyright
+ * ownership. Camunda licenses this file to you under the Apache License,
+ * Version 2.0; you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.camunda.bpm.engine.test.junit5.registerExtension;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
@@ -6,26 +22,26 @@ import org.camunda.bpm.engine.RuntimeService;
 import org.camunda.bpm.engine.TaskService;
 import org.camunda.bpm.engine.task.Task;
 import org.camunda.bpm.engine.test.Deployment;
-import org.camunda.bpm.extension.junit5.test.ProcessEngineExtension;
+import org.camunda.bpm.engine.test.junit5.ProcessEngineExtension;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
 
 public class RegisterProcessEngineExtensionTest {
-  
+
   @RegisterExtension
   ProcessEngineExtension extension = ProcessEngineExtension.builder().build();
-  
+
   @Test
   @Deployment
   public void registeredExtensionUsageExample() {
     RuntimeService runtimeService = extension.getProcessEngine()
         .getRuntimeService();
     runtimeService.startProcessInstanceByKey("registeredExtensionUsage");
-    
+
     TaskService taskService = extension
         .getProcessEngine()
         .getTaskService();
-    
+
     Task task = taskService
         .createTaskQuery()
         .singleResult();
