--- test-utils/junit5-extension/src/test/java/org/camunda/bpm/extension/junit5/ProcessEngineExtensionParentClass.java
@@ -1,5 +1,0 @@
-package org.camunda.bpm.extension.junit5;
-
-public class ProcessEngineExtensionParentClass extends ProcessEngineExtensionSuperClass {
-
-}
