--- test-utils/junit5-extension/src/test/java/org/camunda/bpm/extension/junit5/ProcessEngineExtensionParentClassDeployment.java
@@ -1,8 +1,0 @@
-package org.camunda.bpm.extension.junit5;
-
-import org.camunda.bpm.engine.test.Deployment;
-
-@Deployment
-public class ProcessEngineExtensionParentClassDeployment {
-
-}
