--- test-utils/junit5-extension/src/test/java/org/camunda/bpm/extension/junit5/ProcessEngineExtensionSuperClass.java
@@ -1,8 +1,0 @@
-package org.camunda.bpm.extension.junit5;
-
-import org.camunda.bpm.engine.test.Deployment;
-
-@Deployment(resources = {"processes/superProcess.bpmn", "processes/subProcess.bpmn"})
-public class ProcessEngineExtensionSuperClass {
-
-}
