+++ test-utils/junit5-extension/src/test/resources/org/camunda/bpm/engine/test/junit5/ProcessEngineExtensionClassDeploymentTest.bpmn
